Refactor to start using getattr_static for python3
Avoids having to use AttrCleaner in many contexts
