+++ bpython/test/test_simpleeval.py
@@ -7,8 +7,6 @@ from bpython.simpleeval import (
     simple_eval,
     evaluate_current_expression,
     EvaluationError,
-    safe_get_attribute,
-    safe_get_attribute_new_style,
 )
 from bpython.test import unittest
 from bpython._py3compat import py3
@@ -187,72 +185,6 @@ class OverriddenMRO(object):
 
 
 member_descriptor = type(Slots.s1)
-
-
-class TestSafeGetAttribute(unittest.TestCase):
-    def test_lookup_on_object(self):
-        a = A()
-        a.x = 1
-        self.assertEqual(safe_get_attribute_new_style(a, "x"), 1)
-        self.assertEqual(safe_get_attribute_new_style(a, "a"), "a")
-        b = B()
-        b.y = 2
-        self.assertEqual(safe_get_attribute_new_style(b, "y"), 2)
-        self.assertEqual(safe_get_attribute_new_style(b, "a"), "a")
-        self.assertEqual(safe_get_attribute_new_style(b, "b"), "b")
-
-    def test_avoid_running_properties(self):
-        p = Property()
-        self.assertEqual(safe_get_attribute_new_style(p, "prop"), Property.prop)
-
-    @unittest.skipIf(py3, "Old-style classes not in Python 3")
-    def test_raises_on_old_style_class(self):
-        class Old:
-            pass
-
-        with self.assertRaises(ValueError):
-            safe_get_attribute_new_style(Old, "asdf")
-
-    def test_lookup_with_slots(self):
-        s = Slots()
-        s.s1 = "s1"
-        self.assertEqual(safe_get_attribute(s, "s1"), "s1")
-        self.assertIsInstance(
-            safe_get_attribute_new_style(s, "s1"), member_descriptor
-        )
-        with self.assertRaises(AttributeError):
-            safe_get_attribute(s, "s2")
-        self.assertIsInstance(
-            safe_get_attribute_new_style(s, "s2"), member_descriptor
-        )
-
-    def test_lookup_on_slots_classes(self):
-        sga = safe_get_attribute
-        s = SlotsSubclass()
-        self.assertIsInstance(sga(Slots, "s1"), member_descriptor)
-        self.assertIsInstance(sga(SlotsSubclass, "s1"), member_descriptor)
-        self.assertIsInstance(sga(SlotsSubclass, "s4"), property)
-        self.assertIsInstance(sga(s, "s4"), property)
-
-    @unittest.skipIf(py3, "Py 3 doesn't allow slots and prop in same class")
-    def test_lookup_with_property_and_slots(self):
-        sga = safe_get_attribute
-        s = SlotsSubclass()
-        self.assertIsInstance(sga(Slots, "s3"), property)
-        self.assertEqual(safe_get_attribute(s, "s3"), Slots.__dict__["s3"])
-        self.assertIsInstance(sga(SlotsSubclass, "s3"), property)
-
-    def test_lookup_on_overridden_methods(self):
-        sga = safe_get_attribute
-        self.assertEqual(sga(OverriddenGetattr(), "a"), 1)
-        self.assertEqual(sga(OverriddenGetattribute(), "a"), 1)
-        self.assertEqual(sga(OverriddenMRO(), "a"), 1)
-        with self.assertRaises(AttributeError):
-            sga(OverriddenGetattr(), "b")
-        with self.assertRaises(AttributeError):
-            sga(OverriddenGetattribute(), "b")
-        with self.assertRaises(AttributeError):
-            sga(OverriddenMRO(), "b")
 
 
 if __name__ == "__main__":
